apache-airflow/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-compat==1.1.0
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.1
apache-airflow-providers-fab==1.2.0
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-google==10.21.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-postgres==5.11.2
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
black==24.8.0
blinker==1.8.2
cachelib==0.9.0
cachetools==5.4.0
cattrs==23.2.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
db-dtypes==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.1
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.1
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gcsfs==2024.6.1
google-ads==25.0.0
google-analytics-admin==0.23.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.61.0
google-cloud-appengine-logging==1.4.5
google-cloud-audit-log==0.2.5
google-cloud-automl==2.13.5
google-cloud-batch==0.17.24
google-cloud-bigquery==3.20.1
google-cloud-bigquery-datatransfer==3.15.5
google-cloud-bigtable==2.25.0
google-cloud-build==3.24.2
google-cloud-compute==1.19.2
google-cloud-container==2.50.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.20.1
google-cloud-dataflow-client==0.8.12
google-cloud-dataform==0.5.11
google-cloud-dataplex==2.2.2
google-cloud-dataproc==5.10.2
google-cloud-dataproc-metastore==1.15.5
google-cloud-dlp==3.21.0
google-cloud-kms==2.24.2
google-cloud-language==2.14.0
google-cloud-logging==3.11.1
google-cloud-memcache==1.9.5
google-cloud-monitoring==2.22.2
google-cloud-orchestration-airflow==1.13.1
google-cloud-os-login==2.14.6
google-cloud-pubsub==2.23.0
google-cloud-redis==2.15.5
google-cloud-resource-manager==1.12.5
google-cloud-run==0.10.8
google-cloud-secret-manager==2.20.2
google-cloud-spanner==3.48.0
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-cloud-storage-transfer==1.11.5
google-cloud-tasks==2.16.5
google-cloud-texttospeech==2.16.5
google-cloud-translate==3.16.0
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-cloud-workflows==1.14.5
google-crc32c==1.5.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.65.0
grpcio-gcp==0.2.2
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
json-merge-patch==0.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==24.12.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pandas==2.1.4
pandas-gbq==0.23.1
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
pluggy==1.5.0
prison==0.2.1
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydata-google-auth==1.8.2
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
rsa==4.9
setproctitle==1.3.3
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
sqlalchemy-bigquery==1.11.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
